move GTBlockStateProperties to the data/block package
