+++ src/main/java/com/gregtechceu/gtceu/api/registry/registrate/provider/GTBlockstateProvider.java
@@ -1,6 +1,6 @@
 package com.gregtechceu.gtceu.api.registry.registrate.provider;
 
-import com.gregtechceu.gtceu.api.block.property.GTBlockStateProperties;
+import com.gregtechceu.gtceu.data.block.GTBlockStateProperties;
 import com.gregtechceu.gtceu.api.machine.MachineDefinition;
 import com.gregtechceu.gtceu.api.machine.RotationState;
 import com.gregtechceu.gtceu.client.util.ExtendedBlockModelRotation;
