+++ src/main/java/com/gregtechceu/gtceu/common/block/RubberLogBlock.java
@@ -1,6 +1,6 @@
 package com.gregtechceu.gtceu.common.block;
 
-import com.gregtechceu.gtceu.api.block.property.GTBlockStateProperties;
+import com.gregtechceu.gtceu.data.block.GTBlockStateProperties;
 import com.gregtechceu.gtceu.data.block.GTBlocks;
 
 import net.minecraft.world.item.context.UseOnContext;
